pip_requirements.txt
asgiref==3.2.10
Django==3.0.8
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
djangorestframework==3.11.0
django-bootstrap-modal-forms==2.0.0
django-colorful==1.3
django-cleanup==5.0.0
django-filter==2.3.0
